Move general definitions into separate header file.
