+++ include/fty_common_db.h
@@ -27,45 +27,4 @@
 
 //  Add your own public definitions here, if you need them
 
-//! Possible error types
-enum errtypes {
-    //! First error should be UNKNOWN as it maps to zero and zero is weird
-    UNKNOWN_ERR,
-    DB_ERR,
-    BAD_INPUT,
-    INTERNAL_ERR,
-    LICENSING_ERR,
-};
-
-//! Constants for database errors
-enum db_err_nos {
-    //! First error should be UNKNOWN as it maps to zero and zero is weird
-    DB_ERROR_UNKNOWN,
-    DB_ERROR_INTERNAL,
-    // Probably should be removed at some point and replaced with bad_input_err
-    DB_ERROR_BADINPUT,
-    DB_ERROR_NOTFOUND,
-    DB_ERROR_NOTIMPLEMENTED,
-    DB_ERROR_DBCORRUPTED,
-    DB_ERROR_NOTHINGINSERTED,
-    DB_ERROR_DELETEFAIL,
-    DB_ERROR_CANTCONNECT,
-};
-
-//! Constants for bad input type of error
-enum bad_input_err {
-    //! First error should be UNKNOWN as it maps to zero and zero is weird
-    BAD_INPUT_UNKNOWN,
-    BAD_INPUT_WRONG_INPUT,
-    BAD_INPUT_OUT_OF_BOUNDS,
-};
-
-//! Constants for internal errors
-enum internal_err {
-    //! First error should be UNKNOWN as it maps to zero and zero is weird
-    INTERNAL_UNKNOWN,
-    INTERNAL_NOT_IMPLEMENTED
-};
-
-
 #endif
