+++ project.xml
@@ -74,6 +74,8 @@
         </use>
     </use>
 
+    <header name = "fty_common_db_defs"> Common data types for DB functions </header>
+
     <class name = "fty_common_db_dbpath" selftest = "0" stable = "1" > Fty database path </class>
     <class name = "fty_common_db_asset" selftest = "1" stable = "1" > Asset filtering function </class>
     <class name = "fty_common_db_asset_delete" selftest = "0" stable = "1" > Functions deleting assets from database. </class>
