+++ src/fty_common_db_classes.h
@@ -39,7 +39,6 @@
 //  Internal API
 
 
-
 //  *** To avoid double-definitions, only define if building without draft ***
 #ifndef FTY_COMMON_DB_BUILD_DRAFT_API
